修改文件

+++ src/main/java/org/sq/zbnss/controller/CheckController.java
@@ -1,16 +1,24 @@
 package org.sq.zbnss.controller;
 
 import com.baomidou.mybatisplus.core.metadata.IPage;
+import com.github.xiaoymin.knife4j.annotations.ApiOperationSupport;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.sq.zbnss.base.PageResultVo;
 import org.sq.zbnss.base.ResponseVo;
 import org.sq.zbnss.entity.Check;
-import org.sq.zbnss.service.CheckService;
+import org.sq.zbnss.entity.Dic;
+import org.sq.zbnss.entity.Log;
+import org.sq.zbnss.entity.User;
+import org.sq.zbnss.service.*;
 import org.springframework.web.bind.annotation.*;
+import org.sq.zbnss.uitl.JWTUtil;
 import org.sq.zbnss.uitl.ResultUtil;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpServletRequest;
+import java.text.MessageFormat;
+import java.util.ArrayList;
 
 /**
  * 检查结果记录(TbCheck)表控制层
@@ -28,6 +36,14 @@ public class CheckController {
     @Resource
     private CheckService tbCheckService;
 
+    @Resource
+    private LogService logService;
+
+    @Resource
+    private UserService userService;
+
+    private DicService dicService;
+
     /**
      * 分页查询
      *
@@ -35,8 +51,9 @@ public class CheckController {
      * @return 查询结果
      */
     @ApiOperation(value = "获取检查列表", tags = "检查管理")
+    @ApiOperationSupport(includeParameters = {"systemId.id","checkUser.id","companyId.id","detail","planDate"})
     @GetMapping("/list")
-    public PageResultVo queryByPage(Check tbCheck,
+    public PageResultVo queryByPage( Check tbCheck,
                                     @RequestParam( value = "pageNumber" ,required = true) Integer pageNumber,
                                     @RequestParam(value = "pageSize" ,required = true)Integer pageSize) {
         IPage<Check> checkPage = tbCheckService.queryByPage(tbCheck,pageNumber,pageSize);
@@ -63,14 +80,25 @@ public class CheckController {
      */
     @ApiOperation(value = "新增检查信息", tags = "检查管理")
     @PostMapping("/add")
-    public ResponseVo add(Check tbCheck) {
+    public ResponseVo add(@RequestBody Check tbCheck, HttpServletRequest request) {
         if(tbCheck == null || tbCheck.getCompanyId() == null || tbCheck.getPlanDate() == null){
             return ResultUtil.error("检查单位CompanyId()和计划检查时间（PlanDate）不能为空");
         }
-            Check check = this.tbCheckService.insert(tbCheck);
+
+        ArrayList<Check> old = this.tbCheckService.queryByPage(tbCheck);
+        if(old.size() > 0){
+            return ResultUtil.error("检查信息已存在");
+        }
+        Check check = this.tbCheckService.insert(tbCheck);
+
         if(check == null){
             return ResultUtil.error("检查信息添加失败");
         }
+        User loginUser = getLoginUser(request);
+        Log log = new Log();
+        log.setUserId(loginUser);
+        log.setOperate(MessageFormat.format("新增检查信息{0}",check.toString()));
+        logService.insert(log);
         return ResultUtil.success("检查信息添加成功",check);
     }
 
@@ -82,14 +110,43 @@ public class CheckController {
      */
     @ApiOperation(value = "修改检查信息", tags = "检查管理")
     @PutMapping("/update")
-    public ResponseVo edit(Check tbCheck) {
+    public ResponseVo edit(@RequestBody  Check tbCheck, HttpServletRequest request) {
         if(tbCheck == null || tbCheck.getId() == 0 || tbCheck.getCheckId() == null || "".equals(tbCheck.getCheckId())){
             return ResultUtil.error("请求参数错误");
+        }
+        if(tbCheck.getId() <= 0){
+            return ResultUtil.error("不存在id<=0的检查信息");
+        }
+        Check check_old = this.tbCheckService.queryById(tbCheck.getId());
+        if(check_old == null){
+            return ResultUtil.error("检查信息不存在");
+        }
+        int statusId = tbCheck.getStatus().getId();
+        if(statusId > 0){
+            ArrayList<Dic> statuses = dicService.queryByType(5);
+            boolean flag = false;
+            for(Dic dic : statuses){
+                if(statusId == dic.getId()){
+                    flag = true;
+                    break;
+                }
+            }
+            if(!flag){
+                return ResultUtil.error("检查信息状态只能修改为字典中规定的状态");
+            }
+        }
+        if(check_old.getStatus().getId() == 15 || check_old.getStatus().getId() == 16){
+            return ResultUtil.error("检查已完成，不能修改");
         }
         Check check= this.tbCheckService.update(tbCheck);
         if(check == null){
             return ResultUtil.error("检查信息修改失败");
         }else{
+            User loginUser = getLoginUser(request);
+            Log log = new Log();
+            log.setUserId(loginUser);
+            log.setOperate(MessageFormat.format("修改检查信息{0}",check.toString()));
+            logService.insert(log);
             return ResultUtil.success("检查信息修改成功",check);
         }
     }
@@ -102,17 +159,34 @@ public class CheckController {
      */
     @ApiOperation(value = "删除检查信息", tags = "检查管理")
     @DeleteMapping("/delete")
-    public ResponseVo deleteById(Integer id) {
+    public ResponseVo deleteById(Integer id, HttpServletRequest request) {
         if(id == 0){
             return ResultUtil.error("请求参数错误");
         }
+        Check check_old = this.tbCheckService.queryById(id);
+        if(check_old == null){
+            return ResultUtil.error("检查信息不存在");
+        }
         boolean result = this.tbCheckService.deleteById(id);
         if(result){
+            User loginUser = getLoginUser(request);
+            Log log = new Log();
+            log.setUserId(loginUser);
+            log.setOperate(MessageFormat.format("修改检查信息{0}",check_old.toString()));
+            logService.insert(log);
             return ResultUtil.error("检查信息删除成功");
         }else{
             return ResultUtil.error("检查信息删除失败");
         }
     }
 
+    public User getLoginUser(HttpServletRequest request){
+        String tokenHeader = request.getHeader("Authorization");
+        String[] tokens = tokenHeader.split(" ");
+        String token = tokens[1];
+        String userId = JWTUtil.getUserId(token);
+        User loginUser = userService.getUserByUserId(userId);
+        return loginUser;
+    }
 }
